==> experiments/00_prune_data.ipynb <==
import sys
sys.path.append('..')

from pathlib import Path 
import pandas as pd
from tqdm.notebook import tqdm

DATA = Path('..') / 'data'
RAW = DATA / '00_raw'
PRUNED = DATA / '01_pruned'

CHUNK = 1_000_000

## Plants_gbif_idigbio

in_file = RAW / 'Plants_gbif_idigbio.csv.gz'
out_file = PRUNED / 'Plants_gbif_idigbio.csv.gz'

headers = !zcat "$in_file" | head -1
headers = headers[0].split('\t')
headers

df = pd.read_csv(
    in_file, sep='\t', dtype=str, keep_default_na=False,
    usecols=['gbifID',
             'decimalLatitude', 'decimalLongitude',
             'coordinateUncertaintyInMeters',
             'locality',
             'stateProvince', 'countryCode'])

df = df.rename(columns={
    'gbifID': 'gbifid',
    'decimalLatitude': 'lat',
    'decimalLongitude': 'lng',
    'coordinateUncertaintyInMeters': 'uncert',
    'locality': 'locality',
    'countryCode': 'country',
    'stateProvince': 'state',
})

df.head()

df.to_csv(out_file, index=False)

## insects_idigbio/occurrence_raw

in_file = RAW / 'insects_idigbio' / 'occurrence_raw.csv.gz'
out_file = PRUNED / 'insects_idigbio_occurrence_raw.csv.gz'

headers = !zcat "$in_file" | head -1
headers = headers[0].split(',')
headers

df = pd.read_csv(
    in_file, dtype=str, keep_default_na=False,
    usecols=['coreid',
             'dwc:decimalLatitude', 'dwc:decimalLongitude',
             'dwc:geodeticDatum', 'dwc:coordinateUncertaintyInMeters',
             'dwc:locality', 'dwc:verbatimLocality',
             'dwc:countryCode', 'dwc:stateProvince', 'dwc:county',
            ])

df = df.rename(columns={
    'dwc:decimalLatitude': 'lat',
    'dwc:decimalLongitude': 'lng',
    'dwc:geodeticDatum': 'datum',
    'dwc:coordinateUncertaintyInMeters': 'uncert',
    'dwc:locality': 'locality',
    'dwc:verbatimLocality': 'locality2',
    'dwc:countryCode': 'country',
    'dwc:stateProvince': 'state',
    'dwc:county': 'county',
})

df.head()

df.to_csv(out_file, index=False)

## insects_gbif

in_file = RAW / 'insects_gbif.csv.gz'
out_file = PRUNED / 'insects_gbif.csv.gz'

headers = !zcat "$in_file" | head -1
headers = headers[0].split('\t')
headers

reader = pd.read_csv(
    in_file, sep='\t', dtype=str, keep_default_na=False,
    chunksize=CHUNK,
    usecols=['gbifID',
             'decimalLatitude', 'decimalLongitude',
             'coordinateUncertaintyInMeters',
             'locality',
             'countryCode', 'stateProvince'])

first_chunk = True
for df in tqdm(reader):
    df = df.rename(columns={
        'gbifID': 'gbifid',
        'decimalLatitude': 'lat',
        'decimalLongitude': 'lng',
        'coordinateUncertaintyInMeters': 'uncert',
        'locality': 'locality',
        'countryCode': 'country',
        'stateProvince': 'state',
    })

    if first_chunk:
        df.to_csv(out_file, index=False)
    else:
        df.to_csv(out_file, index=False, mode='a', header=False)
    first_chunk = False

## plants_gbif/verbatim

in_file = RAW / 'plants_gbif' / 'verbatim.txt.gz'
out_file = PRUNED / 'plants_gbif_verbatim.csv.gz'

headers = !zcat "$in_file" | head -1
headers = headers[0].split('\t')
headers

reader = pd.read_csv(
    in_file, sep='\t', dtype=str, keep_default_na=False,
    chunksize=CHUNK,
    usecols=['gbifID',
             'decimalLatitude', 'decimalLongitude',
             'geodeticDatum',
             'coordinateUncertaintyInMeters',
             'locality', 'verbatimLocality',
             'countryCode', 'stateProvince', 'county',
            ])

first_chunk = True
for df in tqdm(reader):
    df = df.rename(columns={
        'gbifID': 'gbifid',
        'decimalLatitude': 'lat',
        'decimalLongitude': 'lng',
        'geodeticDatum': 'datum',
        'coordinateUncertaintyInMeters': 'uncert',
        'locality': 'locality',
        'verbatimLocality': 'locality2',
        'countryCode': 'country',
        'stateProvince': 'state',
    })

    if first_chunk:
        df.to_csv(out_file, index=False)
    else:
        df.to_csv(out_file, index=False, mode='a', header=False)
    first_chunk = False

## plants_idigbio/occurrence_raw

in_file = RAW / 'plants_idigbio' / 'occurrence_raw.csv.gz'
out_file = PRUNED / 'plants_idigbio_occurrence_raw.csv.gz'

headers = !zcat "$in_file" | head -1
headers = headers[0].split(',')
headers

reader = pd.read_csv(
    in_file, dtype=str, keep_default_na=False,
    chunksize=CHUNK,
    usecols=['coreid',
             'dwc:decimalLatitude', 'dwc:decimalLongitude',
             'dwc:geodeticDatum',
             'dwc:coordinateUncertaintyInMeters',
             'dwc:locality', 'dwc:verbatimLocality',
             'dwc:countryCode', 'dwc:stateProvince', 'dwc:county',
            ])

first_chunk = True
for df in tqdm(reader):
    df = df.rename(columns={
        'dwc:decimalLatitude': 'lat',
        'dwc:decimalLongitude': 'lng',
        'dwc:geodeticDatum': 'datum',
        'dwc:coordinateUncertaintyInMeters': 'uncert',
        'dwc:locality': 'locality',
        'dwc:verbatimLocality': 'locality2',
        'dwc:countryCode': 'country',
        'dwc:stateProvince': 'state',
        'dwc:county': 'county',
    })
    if first_chunk:
        df.to_csv(out_file, index=False)
    else:
        df.to_csv(out_file, index=False, mode='a', header=False)
    first_chunk = False

## mammals_gbif/verbatim

in_file = RAW / 'mammals_gbif' / 'verbatim.txt.gz'
out_file = PRUNED / 'mammals_gbif_verbatim.csv.gz'

headers = !zcat "$in_file" | head -1
headers = headers[0].split('\t')
headers

df = pd.read_csv(
    in_file, sep='\t', dtype=str, keep_default_na=False,
    usecols=['gbifID',
             'decimalLatitude', 'decimalLongitude',
             'geodeticDatum',
             'coordinateUncertaintyInMeters',
             'locality', 'verbatimLocality',
             'countryCode', 'stateProvince', 'county',
            ])

df = df.rename(columns={
    'gbifID': 'gbifid',
    'decimalLatitude': 'lat',
    'decimalLongitude': 'lng',
    'geodeticDatum': 'datum',
    'coordinateUncertaintyInMeters': 'uncert',
    'locality': 'locality',
    'verbatimLocality': 'locality2',
    'countryCode': 'country',
    'stateProvince': 'state',
})

df.head()

df.to_csv(out_file, index=False)

## mammals_idigbio

in_file = RAW / 'mammals_idigbio' / 'occurrence_raw.csv.gz'
out_file = PRUNED / 'mammals_idigbio_occurrence_raw.csv.gz'

headers = !zcat "$in_file" | head -1
headers = headers[0].split(',')
headers

df = pd.read_csv(
    in_file, dtype=str, keep_default_na=False,
    usecols=['coreid',
             'dwc:decimalLatitude', 'dwc:decimalLongitude',
             'dwc:geodeticDatum',
             'dwc:coordinateUncertaintyInMeters',
             'dwc:locality', 'dwc:verbatimLocality',
             'dwc:countryCode', 'dwc:stateProvince', 'dwc:county',
            ])

df = df.rename(columns={
    'dwc:decimalLatitude': 'lat',
    'dwc:decimalLongitude': 'lng',
    'dwc:geodeticDatum': 'datum',
    'dwc:coordinateUncertaintyInMeters': 'uncert',
    'dwc:locality': 'locality',
    'dwc:verbatimLocality': 'locality2',
    'dwc:countryCode': 'country',
    'dwc:stateProvince': 'state',
    'dwc:county': 'county',
})

df.head()

df.to_csv(out_file, index=False)
