==> experiments/01_iDigBio_build.ipynb <==
# Build the Gazetteer from iDigBio Data

This is a dump of all of iDigBio data retrieved on 2020-03-30. (No DOI available.)

## Basic Notebook Setup

import os
import io
import sqlite3
import zipfile
from pathlib import Path

import dask.dataframe as dd
import pandas as pd
import regex
from tqdm import tqdm

DATA_DIR = Path('..') / 'data'
RAW_DIR = DATA_DIR / '00_raw'
INTERIM_DIR = DATA_DIR / '01_interim'

ZIP_FILE = RAW_DIR / 'idigbio_all_2020-03-30.zip'
OUT_DB = INTERIM_DIR / 'gazetteer_01_idigbio_2020-03-30.db'

CHUNK = 1_000_000

## Build Gazetteer Related Fields

Gather all of the fields below. Not all of them will make it into the Gazetteer table but we want them all geography related fields available if we tweak the algorithm later.

iDigBio uses name spaces.

DWC_FIELDS = {f'dwc:{f}': f for f in """
    continent
    coordinatePrecision
    coordinateUncertaintyInMeters
    country
    countryCode
    county
    decimalLatitude
    decimalLongitude
    footprintSRS
    footprintSpatialFit
    footprintWKT
    georeferencedDate
    geodeticDatum
    georeferencedBy
    georeferenceProtocol
    georeferenceremarks
    georeferenceSources
    georeferenceVerificationStatus
    higherGeography
    higherGeographyID
    island
    islandGroup
    locality
    locationID
    locationRemarks
    maximumDepthInMeters
    maximumDistanceAboveSurfaceInMeters
    maximumElevationInMeters
    minimumDepthInMeters
    minimumDistanceAboveSurfaceInMeters
    minimumElevationInMeters
    municipality
    pointRadiusSpatialFit
    stateProvince
    verbatimCoordinateSystem
    verbatimCoordinates
    verbatimDepth
    verbatimElevation
    verbatimLatitude
    verbatimLocality
    verbatimLongitude
    verbatimSRS
    waterBody
""".split()}


IDIGBIO_FIELDS = {f'idigbio:{f}': f for f in """
    geoPoint
    isoCountryCode
""".split()}

OTHER_FIELDS = {'coreid': 'coreid'}  # Need to link data

FIELDS = {**OTHER_FIELDS, **DWC_FIELDS, **IDIGBIO_FIELDS}

## Examine Tables

What's in the downloaded zip file and what is in the tables.

with zipfile.ZipFile(ZIP_FILE) as zippy:
    names = zippy.namelist()

names

The occurrence.csv & occurrence_raw.csv files may have what we need. Look at their headers.

### occurrence.csv

with zipfile.ZipFile(ZIP_FILE) as zippy:
    with zippy.open('occurrence.csv') as in_file:
        headers = in_file.readline()

headers = [h.decode().strip() for h in sorted(headers.split(b','))]

#### All Headers in occurrence.csv

sorted(headers)

#### Target Headers in occurrence.csv

HEADERS = sorted(h for h in headers if h in FIELDS)
HEADERS

### occurrence_raw.csv

with zipfile.ZipFile(ZIP_FILE) as zippy:
    with zippy.open('occurrence_raw.csv') as in_file:
        headers = in_file.readline()

headers = [h.decode().strip() for h in sorted(headers.split(b','))]

#### All headers in occurrence_raw.csv

sorted(headers)

#### Target headers in occurrence_raw.csv

RAW_HEADERS = sorted(h for h in headers if h in FIELDS)

## Remove Old Database

if OUT_DB.exists():
    os.remove(OUT_DB)

## Create Input Tables

Grab all of the relevant fields from the table if the record has a valid latitude and longitude.

### Create occurrence Table

renames = {h: FIELDS[h] for h in HEADERS}

with sqlite3.connect(OUT_DB) as cxn:
    with zipfile.ZipFile(ZIP_FILE) as zippy:
        with zippy.open('occurrence.csv') as in_file:
            reader = pd.read_csv(
                in_file, dtype=str, keep_default_na=False,
                chunksize=CHUNK, usecols=HEADERS)

            if_exists = 'replace'

            for df in tqdm(reader):
                df = df.rename(columns=renames)

                has_point = df['geoPoint'] != ''

                df.loc[has_point, :].to_sql(
                    'occurrence', cxn, if_exists=if_exists, index=False)

                if_exists = 'append'

### Create occurrence_raw Table

renames = {h: FIELDS[h] for h in RAW_HEADERS}

with sqlite3.connect(OUT_DB) as cxn:
    with zipfile.ZipFile(ZIP_FILE) as zippy:
        with zippy.open('occurrence_raw.csv') as in_file:
            reader = pd.read_csv(
                in_file, dtype=str, keep_default_na=False,
                chunksize=CHUNK, usecols=RAW_HEADERS)

            if_exists = 'replace'

            for df in tqdm(reader):
                df = df.rename(columns=renames)

                df['decimalLatitude'] = pd.to_numeric(
                    df['decimalLatitude'], errors='coerce')
                df['decimalLongitude'] = pd.to_numeric(
                    df['decimalLongitude'], errors='coerce')

                has_lat = df['decimalLatitude'].notna()
                has_lng = df['decimalLongitude'].notna()

                df.loc[has_lat & has_lng, :].to_sql(
                    'occurrence_raw', cxn, if_exists=if_exists, index=False)

                if_exists = 'append'
