==> experiments/02_unique_places.ipynb <==
import sys
sys.path.append('..')

import os
from pathlib import Path
import sqlite3
import regex
import numpy as np
import pandas as pd
import dask.dataframe as dd

DATA = Path('..') / 'data'
PRUNED = DATA / '01_pruned'
INPUT = DATA / 'input'

RAW = PRUNED / 'raw.csv'
DB = PRUNED / 'gazetteer.db'

OUTPUT = PRUNED / 'unique.csv'
TEMP_DIR = PRUNED / 'unique.csv.temp'
TEMP_FILE = TEMP_DIR / '0.part'

TARGET = INPUT / 'mammals_no_geo.csv.gz'

df = dd.read_csv(RAW, dtype=str).fillna('')
df

df = df.drop(columns=['gbifid', 'coreid', 'country', 'state', 'county'])

## Normalize locality values

- Space normalize
- Lower case
- Remove punctuation except for "." decimal points, dots surrounded by digits

remove = r'(?<!\d)[.,;/(){}"\'\[\]\-](?!\d)'

df['norm'] = df['locality'].str.replace(remove, ' ')
df['norm'] = df['norm'].str.lower().str.split().str.join(' ')

## Need to drop AFTER other manipulations because the drop reduces to a single thread

df = df.drop_duplicates()
df

df.to_csv(TEMP_DIR, index=False)

## Rename output file

os.rename(TEMP_FILE, OUTPUT)
os.rmdir(TEMP_DIR)

## Update data types

df = pd.read_csv(
    str(OUTPUT),
    dtype={'locality': str,
           'lat': np.float32,
           'lng': np.float32,
           'dataum': str,
           'uncert': np.int32,
           })

with sqlite3.connect(DB) as cxn:
    cxn.row_factory = sqlite3.Row
    df.to_sql('places', cxn, if_exists='replace', index=False)

with sqlite3.connect(DB) as cxn:
    sql = """
        CREATE INDEX places_locality ON places (locality);
        CREATE INDEX places_norm ON places (norm);
        """
    cxn.executescript(sql)
