==> experiments/10_idigbio.ipynb <==
import sys
sys.path.append('..')

import os
from pathlib import Path
import sqlite3
import zipfile
import regex
import numpy as np
import pandas as pd
import dask.dataframe as dd

DATA_DIR = Path('..') / 'data'
RAW_DIR = DATA_DIR / '00_raw'
INTERIM_DIR = DATA_DIR / '01_interim'

ZIP_FILE = RAW_DIR / 'idigbio_all.zip'

with zipfile.ZipFile(ZIP_FILE) as zippy:
    with zippy.open('occurrence_raw.csv') as in_file:
        occurrence_raw = in_file.readline()
occurrence_raw = [h.decode().strip() for h in sorted(occurrence_raw.split(b','))]
occurrence_raw

with zipfile.ZipFile(ZIP_FILE) as zippy:
    with zippy.open('occurrence.csv') as in_file:
        occurrence = in_file.readline()
occurrence = [h.decode().strip() for h in sorted(occurrence.split(b','))]
occurrence

fields = [f'dwc:{f}' for f in """
    locationID higherGeographyID higherGeography
    continent
    waterBody islandGroup island
    country countryCode stateProvince county municipality
    locality verbatimLocality 
    minimumElevationInMeters maximumElevationInMeters
    verbatimElevation 
    minimumDepthInMeters maximumDepthInMeters verbatimDepth 
    minimumDistanceAboveSurfaceInMeters
    maximumDistanceAboveSurfaceInMeters 
    locationAccordingTo locationRemarks 
    decimalLatitude decimalLongitude geodeticDatum 
    coordinateUncertaintyInMeters coordinatePrecision
    pointRadiusSpatialFit 
    verbatimCoordinates verbatimLatitude verbatimLongitude 
    verbatimCoordinateSystem verbatimSRS 
    footprintWKT footprintSRS footprintSpatialFit 
    georeferencedBy georeferencedDate georeferenceProtocol 
    georeferenceSources georeferenceVerificationStatus 
    georeferenceRemarks
""".split()] + [f'idigbio:{f}' for f in """
    isoCountryCode geoPoint
""".split()]

fields

[f for f in occurrence if f in fields]

[f for f in occurrence_raw if f in fields]
